move extraneous factory methods to point of use
