+++ pkg/oc/cli/cmd/set/env.go
@@ -21,6 +21,7 @@ import (
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	"k8s.io/kubernetes/pkg/kubectl/resource"
 
+	buildapi "github.com/openshift/origin/pkg/build/apis/build"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
 	"github.com/openshift/origin/pkg/oc/cli/util/clientcmd"
 	utilenv "github.com/openshift/origin/pkg/oc/util/env"
@@ -406,7 +407,7 @@ func (o *EnvOptions) RunEnv(f kcmdutil.Factory) error {
 		}))
 		if !ok {
 			// This is a fallback function for objects that don't have pod spec.
-			ok, err = f.UpdateObjectEnvironment(info.Object, func(vars *[]kapi.EnvVar) error {
+			ok, err = updateObjectEnvironment(info.Object, func(vars *[]kapi.EnvVar) error {
 				if vars == nil {
 					return fmt.Errorf("no environment variables provided")
 				}
@@ -494,4 +495,24 @@ updates:
 		return kcmdutil.ErrExit
 	}
 	return nil
+}
+
+// UpdateObjectEnvironment update the environment variables in object specification.
+func updateObjectEnvironment(obj runtime.Object, fn func(*[]kapi.EnvVar) error) (bool, error) {
+	switch t := obj.(type) {
+	case *buildapi.BuildConfig:
+		if t.Spec.Strategy.CustomStrategy != nil {
+			return true, fn(&t.Spec.Strategy.CustomStrategy.Env)
+		}
+		if t.Spec.Strategy.SourceStrategy != nil {
+			return true, fn(&t.Spec.Strategy.SourceStrategy.Env)
+		}
+		if t.Spec.Strategy.DockerStrategy != nil {
+			return true, fn(&t.Spec.Strategy.DockerStrategy.Env)
+		}
+		if t.Spec.Strategy.JenkinsPipelineStrategy != nil {
+			return true, fn(&t.Spec.Strategy.JenkinsPipelineStrategy.Env)
+		}
+	}
+	return false, fmt.Errorf("object does not contain any environment variables")
 }
